+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add Icon (#19)
 - Adding compodoc (#23)
 - Adding changelog information to the Contribution.md (#41)
+- Add version-script to obtain package.json information (#35)
 - Add import for csv-files (#39)
 Formatting needs to be similar to the dataformat of the old OmnAIView data exports. For more infos, see comments in code.
 - Add axis mode: Allow to switch between absolute and relative x-axis timestamps(#58)
@@ -45,4 +46,6 @@ Formatting needs to be similar to the dataformat of the old OmnAIView data expor
 
 - Deletion of deprecated Angular 18 patterns (#18)
 - Deletion of duplicated AsyncAPI description for OmnAIBackend 
+- BREAKING CHANGE: Remove `package.json` build scripts in subfolders (#35)
+